factor helper functions into own class for better modularity
 When refactoring the main app into different flask blueprints I will have to have access to these methods from a central location. This static helper class will enable me to do this. The name is still quite stupid. I will have to think about it in more detail.
